+++ gluecodium/src/test/resources/smoke/lambdas/output/android/com/example/smoke/StandaloneProducer.java
@@ -1,0 +1,8 @@
+/*
+ *
+ */
+package com.example.smoke;
+@FunctionalInterface
+public interface StandaloneProducer {
+    String apply();
+}