+++ gluecodium/src/test/resources/smoke/lambdas/output/cpp/include/smoke/StandaloneProducer.h
@@ -1,0 +1,11 @@
+// -------------------------------------------------------------------------------------------------
+//
+//
+// -------------------------------------------------------------------------------------------------
+#pragma once
+#include "gluecodium/Export.h"
+#include <functional>
+#include <string>
+namespace smoke {
+using StandaloneProducer = ::std::function<::std::string()>;
+}